refactor: AddStamp -> AddableStampList
