+++ src/components/Settings/StampPaletteTab/StampPaletteEditorAddableStampListItem.vue
